Migrate to Bits8.

+++ bytes.h
@@ -1,6 +1,7 @@
 #ifndef BYTES_H
 #define BYTES_H
 
+#include <stdint.h>
 #include <stddef.h>
 
 /**
@@ -15,15 +16,15 @@ typedef struct Slice Slice;
 /// Allocate an empty slice with the given capacity.
 Slice * bytes_alloc(size_t capacity);
 
-Slice * bytes_snoc(Slice * slice, unsigned byte);
+Slice * bytes_snoc(Slice * slice, uint8_t byte);
 
 // O(1).
 size_t bytes_length(Slice * slice);
 
 /// Generic O(1) indexing. Does not perform any checks.
-unsigned bytes_at(Slice * slice, size_t index);
-unsigned bytes_head(Slice * slice);
-unsigned bytes_last(Slice * slice);
+uint8_t bytes_at(Slice * slice, size_t index);
+uint8_t bytes_head(Slice * slice);
+uint8_t bytes_last(Slice * slice);
 
 /// Get a subslice corresponding to the half-open interval [start, end).
 /// Does not perform any checks.
